+++ Packages/net.shivaduke28.tauburn/Editor/Animation/AnimationUIBuilderEditor.cs
@@ -14,7 +14,7 @@ namespace Tauburn.Editor.Animation
             if (GUILayout.Button("Create Animation UI"))
             {
                 var uiBuilder = (AnimationUIBuilder) target;
-                UIBuilder.BuildAnimationUI(uiBuilder.LayoutUIMarker, uiBuilder.AnimationBuildDatum, uiBuilder.transform);
+                AnimatorBridgeBuilder.Build(uiBuilder.AnimationBuildDatum, uiBuilder.LayoutUIMarker);
             }
         }
     }
